🎉 min check
